requirements.txt
pandas>=1.5.0,<3.0.0
numpy>=1.23
streamlit>=1.24
plotly>=5.14
PyYAML>=6.0
lxml>=4.9
rapidfuzz>=3.0
scikit-learn>=1.2
Jinja2>=3.1
azure-functions>=1.17
azure-storage-blob>=12.19
azure-identity>=1.15
